requirements.txt
openai == 1.33.0
streamlit == 1.35.0
python-dotenv == 1.0.1
